modificando los archivos 1
Modificando los archivos de la P5 iniciales para que tengan más sentido
