+++ .gitignore
@@ -12,3 +12,5 @@
 /practica5/practica4/build/
 /practica5/practica4/dist/
 /practica5/practica4/nbproject/private/
+/practica5/practica5/build/
+/practica5/practica5/dist/
