+++ practica4/practica4/nbproject/private/private.xml
@@ -2,10 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/Franelas/AppData/Local/GitHubDesktop/app-1.0.3/Practicas-RI/practica4/practica4/src/practica4/FiltraLetras.java</file>
-            <file>file:/C:/Users/Franelas/AppData/Local/GitHubDesktop/app-1.0.3/Practicas-RI/practica4/practica4/src/practica4/Practica4.java</file>
-            <file>file:/C:/Users/Franelas/AppData/Local/GitHubDesktop/app-1.0.3/Practicas-RI/practica4/practica4/src/practica4/Analizador.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
