+++ practica5/practica5/nbproject/genfiles.properties
@@ -1,8 +1,8 @@
-build.xml.data.CRC32=638ed090
-build.xml.script.CRC32=8396ce08
+build.xml.data.CRC32=6b9d5cb6
+build.xml.script.CRC32=5c539cbb
 build.xml.stylesheet.CRC32=8064a381@1.80.1.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=638ed090
-nbproject/build-impl.xml.script.CRC32=c986c039
+nbproject/build-impl.xml.data.CRC32=6b9d5cb6
+nbproject/build-impl.xml.script.CRC32=b56fdd38
 nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
